+++ packages/url_launcher/url_launcher_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.1
+
+- Added support for mailto scheme
+
 # 0.1.0+2
 
 - Remove androidx references from the no-op android implemenation.