+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,6 +1,11 @@
-## 0.10.7
+## 0.10.8
 
 * Add caching functionality to videos from network sources.
+
+## 0.10.7
+
+* `VideoPlayerController` support for reading closed caption files. 
+* `VideoPlayerValue` has a `caption` field for reading the current closed caption at any given time.
 
 ## 0.10.6